Update image paths in documentation
The paths to images in multiple markdown files have been corrected. This ensures that all necessary images are correctly displayed in the CompressorOperation and ClearanceOfVessels pages in the docs. Additionally, a note has been added to the CompressorOperation document reminding users to check each tank's Hydro and Visual Inspection status.